Fixup: Move model folder into app
- Also move errors into models
