+++ internal/app/model/errors.go
@@ -1,4 +1,4 @@
-package app
+package model
 
 import "errors"
