+++ internal/app/store/sqlStore/sqlStore_test/userRepository_test.go
@@ -1,9 +1,8 @@
 package sqlStore_test
 
 import (
-	"github.com/Htomsik/GO-REST-API-Sample/internal/app"
+	"github.com/Htomsik/GO-REST-API-Sample/internal/app/model"
 	"github.com/Htomsik/GO-REST-API-Sample/internal/app/store/sqlStore"
-	"github.com/Htomsik/GO-REST-API-Sample/internal/model"
 	"github.com/stretchr/testify/assert"
 	"testing"
 )
@@ -33,7 +32,7 @@ func TestUserRepository_FindByEmailRandom(t *testing.T) {
 	_, err := st.User().FindByEmail("user@ex.com")
 
 	// Assert
-	assert.EqualError(t, err, app.RecordNotFound.Error())
+	assert.EqualError(t, err, model.RecordNotFound.Error())
 }
 
 func TestUserRepository_FindByEmailAdded(t *testing.T) {
