+++ internal/app/store/testStore/userRepository.go
@@ -1,8 +1,7 @@
 package testStore
 
 import (
-	"github.com/Htomsik/GO-REST-API-Sample/internal/app"
-	"github.com/Htomsik/GO-REST-API-Sample/internal/model"
+	"github.com/Htomsik/GO-REST-API-Sample/internal/app/model"
 )
 
 // UserRepository ...
@@ -34,7 +33,7 @@ func (repository *UserRepository) FindByEmail(email string) (*model.User, error)
 	user, exist := repository.users[email]
 
 	if !exist {
-		return nil, app.RecordNotFound
+		return nil, model.RecordNotFound
 	}
 
 	return user, nil
